fix some tests and misc behavior

+++ app/ios.js
@@ -1487,14 +1487,23 @@ IOS.prototype.localScreenshot = function(desiredFile, cb) {
   async.series([
     function (cb) { this.proxy(command, cb); }.bind(this),
     function (cb) {
+      var srcFile = filePath + ".png";
+      var waitForFile = function() {
+        if (fs.existsSync(srcFile)) {
+          var desiredFolder = path.dirname(desiredFile);
+          mkdirp.sync(desiredFolder);
+          fs.rename(filePath + ".png", desiredFile, cb);
+        } else {
+          setTimeout(waitForFile, 500);
+        }
+      };
+      waitForFile();
       // must exist or rename will fail.
-      var desiredFolder = path.dirname(desiredFile);
-      mkdirp.sync(desiredFolder);
-      fs.rename(filePath + ".png", desiredFile, cb);
-    }.bind(this),
+    },
   ], function(){
     cb(null, {
        status: status.codes.Success.code
+       , value: true
      });
   });
 };
